apps/api/requirements.txt
fastapi
uvicorn
python-dotenv
requests
yt-dlp
moviepy
jinja2
python-multipart
sqlalchemy
google-auth-oauthlib
python-jose
